+++ instapy2/__init__.py
@@ -1,4 +1,4 @@
-version = '0.0.15'
+version = '0.0.16'
 
 # somehow fixes a stupid relative module import error?
 from .instapy2 import InstaPy2