+++ instapy2/media_product_type.py
@@ -1,0 +1,13 @@
+from enum import Enum
+
+class MediaType(Enum):
+    Album = 8,
+    IGTV = 2,
+    Photo = 1
+    Reel = 2
+    Video = 2
+
+class ProductType(Enum):
+    IGTV = 'igtv'
+    Reel = 'clips'
+    Video = 'feed'