+++ instapy2/media_type.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-class MediaType(Enum):
-    Carousel = 'Carousel',
-    Clip = 'Clip',
-    IGTV = 'IGTV',
-    Photo = 'Photo',
-    Video = 'Video'